Save website state

+++ app/Model/Entity/WebsiteIdentify.php
@@ -10,7 +10,7 @@ class WebsiteIdentify extends Website implements WebsiteIdentifyInterface {
 
 	use EntityIdentity;
 
-	public function __construct(int $id, Url $url, ?DateTime $lastCheckAt = NULL, ?int $responseCode = NULL, ?int $responseTime = NULL) {
+	public function __construct(int $id, Url $url, ?DateTime $lastCheckAt = NULL, ?int $responseCode = NULL, ?float $responseTime = NULL) {
 		parent::__construct($url, $lastCheckAt, $responseCode, $responseTime);
 
 		$this->id = $id;
